Move warm migration implementation from libvmmapi to usr.sbin/bhyve/migration.c
