+++ usr.sbin/bhyve/migration.h
@@ -44,4 +44,30 @@ int init_checkpoint_thread(struct vmctx *ctx);
 
 int load_restore_file(const char *filename, struct restore_state *rstate);
 
+/* Warm Migration */
+
+enum message_types {
+    MESSAGE_TYPE_SPECS	    = 1,
+    MESSAGE_TYPE_METADATA   = 2,
+    MESSAGE_TYPE_RAM	    = 3,
+    MESSAGE_TYPE_KERN	    = 4,
+    MESSAGE_TYPE_PCI	    = 5,
+};
+
+struct __attribute__((packed)) migration_message_type {
+    size_t len;
+    unsigned int type;		// enum message_type
+    unsigned int req_type;	// enum snapshot_req
+};
+
+struct __attribute__((packed)) migration_system_specs {
+	char hw_machine[MAX_SPEC_LEN];
+	char hw_model[MAX_SPEC_LEN];
+	size_t hw_pagesize;
+};
+
+int vm_send_migrate_req(struct vmctx *ctx, struct migrate_req req, void *pci_ptr);
+int vm_recv_migrate_req(struct vmctx *ctx, struct migrate_req req, void *pci_ptr);
+
+
 #endif
